+++ Assets/_SIO/Scripts/Components/Health/IHealthComponent.cs
@@ -3,5 +3,5 @@ public interface IHealthComponent
     float Health { get; }
     float MaxHealth { get; }
 
-    void SetDamage(int damage);
+    void SetDamage(float damage);
 }
